perf: _init_state 移除冗余 .astype(np.float64)
np.diag() 传入 Python float 列表默认已产生 float64 数组，
.astype(np.float64) 是多余的拷贝操作。

+++ aim_pipeline.py
@@ -43,7 +43,7 @@ class KalmanPredictor2D:
     def _init_state(self, x, y):
         """初始化单个 track 的卡尔曼状态"""
         state = np.array([x, y, 0.0, 0.0], dtype=np.float64)
-        P = np.diag([100.0, 100.0, 500.0, 500.0]).astype(np.float64)
+        P = np.diag([100.0, 100.0, 500.0, 500.0])
         return state, P
 
     def _get_F(self, dt):
